+++ lib/pages/cacheImageTablePage.dart
@@ -20,7 +20,7 @@ class _CacheImageTablePageState extends State<CacheImageTablePage> {
     setState(() {
       list = LocalCacheSync.imageCache.all;
     });
-    print(list[0].readAsBytesSync());
+    // print(list[0].readAsBytesSync());
     super.initState();
   }